refactor: modularize task manager scripts
