+++ index.html
@@ -488,6 +488,6 @@
     <script type="module" src="firebase.js"></script>
     <script type="module" src="pwa.js"></script>
     <script type="module" src="auth.js"></script>
-    <script type="module" src="script.js"></script>
+    <script type="module" src="src/main.js"></script>
 </body>
 </html>
